Add 0x12. Advanced Binary Search
Function that searches for a value in a sorted array of integers.

You may have noticed that basic binary search does not necessarily
return the index of the first value in the array (if this value
appears more than once in the array).

In this exercise, you’ll have to solve this problem.